pyproject.toml
[tool.poetry]
name = "ark-noid"
version = "0.1.0"
description = ""
authors = ["mary-crkn <mzheng@crkn.ca>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
uvicorn = "^0.32.0"
sqlalchemy = "^2.0.36"
asyncpg = "^0.30.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
alembic = "^1.14.0"
pysolr = "^3.10.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
